// tiptap-templates/templates/next-block-editor-app/src/components/BlockEditor/types.tsx
//import type { Language } from '@/extensions/Ai'

export type EditorUser = {
  clientId: string
  name: string
  color: string
  initials?: string,
  showName:string,
}

// export type LanguageOption = {
//   name: string
//   label: string
//   value: Language
// }

export type AiTone =
  | 'academic'
  | 'business'
  | 'casual'
  | 'childfriendly'
  | 'conversational'
  | 'emotional'
  | 'humorous'
  | 'informative'
  | 'inspirational'
  | string

export type AiPromptType = 'SHORTEN' | 'EXTEND' | 'SIMPLIFY' | 'TONE'

// export type AiToneOption = {
//   name: string
//   label: string
//   value: AiTone
// }

// export type AiImageStyle = {
//   name: string
//   label: string
//   value: string
// }